Plugins can hook now without change the Vault code
