+++ plugin.yml
@@ -1,5 +1,5 @@
 name: Vault
-main: net.milkbowl.vault.Vault
+main: net.milkbowl.vault.VaultPlugin
 version: 1.2.25-b${BUILD_NUMBER}
 authors: [cereal, Sleaker, mung3r]
 website: http://dev.bukkit.org/server-mods/vault
