+++ src/net/milkbowl/vault/VaultPlugin.java
@@ -1,0 +1,330 @@
+/* This file is part of Vault.
+
+    Vault is free software: you can redistribute it and/or modify
+    it under the terms of the GNU Lesser General Public License as published by
+    the Free Software Foundation, either version 3 of the License, or
+    (at your option) any later version.
+
+    Vault is distributed in the hope that it will be useful,
+    but WITHOUT ANY WARRANTY; without even the implied warranty of
+    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+    GNU Lesser General Public License for more details.
+
+    You should have received a copy of the GNU Lesser General Public License
+    along with Vault.  If not, see <http://www.gnu.org/licenses/>.
+ */
+package net.milkbowl.vault;
+
+import java.io.IOException;
+import java.lang.reflect.InvocationTargetException;
+import java.lang.reflect.Method;
+import java.net.URL;
+import java.util.Collection;
+import java.util.logging.Logger;
+
+import javax.xml.parsers.DocumentBuilderFactory;
+
+import net.milkbowl.vault.chat.Chat;
+import net.milkbowl.vault.economy.Economy;
+import net.milkbowl.vault.permission.Permission;
+
+import org.bukkit.Bukkit;
+import org.bukkit.OfflinePlayer;
+import org.bukkit.command.Command;
+import org.bukkit.command.CommandSender;
+import org.bukkit.entity.Player;
+import org.bukkit.event.EventHandler;
+import org.bukkit.event.EventPriority;
+import org.bukkit.event.Listener;
+import org.bukkit.event.player.PlayerJoinEvent;
+import org.bukkit.event.server.PluginEnableEvent;
+import org.bukkit.plugin.RegisteredServiceProvider;
+import org.bukkit.plugin.ServicesManager;
+import org.bukkit.plugin.java.JavaPlugin;
+import org.w3c.dom.Document;
+import org.w3c.dom.Element;
+import org.w3c.dom.Node;
+import org.w3c.dom.NodeList;
+
+import com.nijikokun.register.payment.Methods;
+
+public class VaultPlugin extends JavaPlugin {
+
+    private static final Logger log = Logger.getLogger("Minecraft");
+    private double newVersion;
+    private double currentVersion;
+    private static ServicesManager sm;
+    private Metrics metrics;
+    private static Vault vault;
+
+    @Override
+    public void onDisable() {
+        // Remove all Service Registrations
+        getServer().getServicesManager().unregisterAll(this);
+        Bukkit.getScheduler().cancelTasks(this);
+    }
+
+    @Override
+    public void onEnable() {
+        currentVersion = Double.valueOf(getDescription().getVersion().split("-")[0].replaceFirst("\\.", ""));
+        sm = getServer().getServicesManager();
+        // Make Vaut Object
+        vault = new Vault(sm, this);
+
+        getCommand("vault-info").setExecutor(this);
+        getCommand("vault-convert").setExecutor(this);
+        getServer().getPluginManager().registerEvents(new VaultListener(), this);
+
+        // Schedule to check the version every 30 minutes for an update. This is to update the most recent 
+        // version so if an admin reconnects they will be warned about newer versions.
+        this.getServer().getScheduler().runTaskTimerAsynchronously(this, new Runnable() {
+
+            @Override
+            public void run() {
+                try {
+                    newVersion = updateCheck(currentVersion);
+                    if (newVersion > currentVersion) {
+                        log.warning("Vault " + newVersion + " is out! You are running: Vault " + currentVersion);
+                        log.warning("Update Vault at: http://dev.bukkit.org/server-mods/vault");
+                    }
+                } catch (Exception e) {
+                    // ignore exceptions
+                }
+            }
+
+        }, 0, 432000);
+
+        // Load up the Plugin metrics
+        try {
+            String authors = "";
+            for (String author : this.getDescription().getAuthors()) {
+                authors += author + ", ";
+            }
+            if (!authors.isEmpty()) {
+                authors = authors.substring(0, authors.length() - 2);
+            }
+            metrics = new Metrics(getDescription().getVersion(), authors);
+            metrics.findCustomData(this);
+            metrics.beginMeasuringPlugin(this);
+        } catch (IOException e) {
+            // ignore exception
+        }
+        log.info(String.format("[%s] Enabled Version %s", getDescription().getName(), getDescription().getVersion()));
+    }
+
+    @Override
+    public boolean onCommand(CommandSender sender, Command command, String commandLabel, String[] args) {
+        if (sender instanceof Player) {
+            // Check if Player
+            // If so, ignore command if player is not Op
+            Player p = (Player) sender;
+            if (!p.isOp()) {
+                return true;
+            }
+        }
+
+        if (command.getName().equalsIgnoreCase("vault-info")) {
+            infoCommand(sender);
+            return true;
+        } else if (command.getName().equalsIgnoreCase("vault-convert")) {
+            convertCommand(sender, args);
+            return true;
+        } else {
+            // Show help
+            sender.sendMessage("Vault Commands:");
+            sender.sendMessage("  /vault-info - Displays information about Vault");
+            sender.sendMessage("  /vault-convert [economy1] [economy2] - Converts from one Economy to another");
+            return true;
+        }
+    }
+
+    private void convertCommand(CommandSender sender, String[] args) {
+        Collection<RegisteredServiceProvider<Economy>> econs = this.getServer().getServicesManager().getRegistrations(Economy.class);
+        if (econs == null || econs.size() < 2) {
+            sender.sendMessage("You must have at least 2 economies loaded to convert.");
+            return;
+        } else if (args.length != 2) {
+            sender.sendMessage("You must specify only the economy to convert from and the economy to convert to. (names should not contain spaces)");
+            return;
+        }
+        Economy econ1 = null;
+        Economy econ2 = null;
+        for (RegisteredServiceProvider<Economy> econ : econs) {
+            String econName = econ.getProvider().getName().replace(" ", "");
+            if (econName.equalsIgnoreCase(args[0])) {
+                econ1 = econ.getProvider();
+            } else if (econName.equalsIgnoreCase(args[1])) {
+                econ2 = econ.getProvider();
+            }
+        }
+
+        if (econ1 == null) {
+            sender.sendMessage("Could not find " + args[0] + " loaded on the server, check your spelling");
+            return;
+        } else if (econ2 == null) {
+            sender.sendMessage("Could not find " + args[1] + " loaded on the server, check your spelling");
+            return;
+        }
+
+        sender.sendMessage("This may take some time to convert, expect server lag.");
+        for (OfflinePlayer op : Bukkit.getServer().getOfflinePlayers()) {
+            String pName = op.getName();
+            if (econ1.hasAccount(pName)) {
+                if (econ2.hasAccount(pName)) {
+                    continue;
+                }
+                econ2.createPlayerAccount(pName);
+                econ2.depositPlayer(pName, econ1.getBalance(pName));
+            }
+        }
+        sender.sendMessage("Converson complete, please verify the data before using it.");
+    }
+
+    private void infoCommand(CommandSender sender) {
+        // Get String of Registered Economy Services
+        String registeredEcons = null;
+        Collection<RegisteredServiceProvider<Economy>> econs = this.getServer().getServicesManager().getRegistrations(Economy.class);
+        for (RegisteredServiceProvider<Economy> econ : econs) {
+            Economy e = econ.getProvider();
+            if (registeredEcons == null) {
+                registeredEcons = e.getName();
+            } else {
+                registeredEcons += ", " + e.getName();
+            }
+        }
+
+        // Get String of Registered Permission Services
+        String registeredPerms = null;
+        Collection<RegisteredServiceProvider<Permission>> perms = this.getServer().getServicesManager().getRegistrations(Permission.class);
+        for (RegisteredServiceProvider<Permission> perm : perms) {
+            Permission p = perm.getProvider();
+            if (registeredPerms == null) {
+                registeredPerms = p.getName();
+            } else {
+                registeredPerms += ", " + p.getName();
+            }
+        }
+
+        String registeredChats = null;
+        Collection<RegisteredServiceProvider<Chat>> chats = this.getServer().getServicesManager().getRegistrations(Chat.class);
+        for (RegisteredServiceProvider<Chat> chat : chats) {
+            Chat c = chat.getProvider();
+            if (registeredChats == null) {
+                registeredChats = c.getName();
+            } else {
+                registeredChats += ", " + c.getName();
+            }
+        }
+
+        // Get Economy & Permission primary Services
+        RegisteredServiceProvider<Economy> rsp = getServer().getServicesManager().getRegistration(Economy.class);
+        Economy econ = null;
+        if (rsp != null) {
+            econ = rsp.getProvider();
+        }
+        Permission perm = null;
+        RegisteredServiceProvider<Permission> rspp = getServer().getServicesManager().getRegistration(Permission.class);
+        if (rspp != null) {
+            perm = rspp.getProvider();
+        }
+        Chat chat = null;
+        RegisteredServiceProvider<Chat> rspc = getServer().getServicesManager().getRegistration(Chat.class);
+        if (rspc != null) {
+            chat = rspc.getProvider();
+        }
+        // Send user some info!
+        sender.sendMessage(String.format("[%s] Vault v%s Information", getDescription().getName(), getDescription().getVersion()));
+        sender.sendMessage(String.format("[%s] Economy: %s [%s]", getDescription().getName(), econ == null ? "None" : econ.getName(), registeredEcons));
+        sender.sendMessage(String.format("[%s] Permission: %s [%s]", getDescription().getName(), perm == null ? "None" : perm.getName(), registeredPerms));
+        sender.sendMessage(String.format("[%s] Chat: %s [%s]", getDescription().getName(), chat == null ? "None" : chat.getName(), registeredChats));
+    }
+
+    /**
+     * Determines if all packages in a String array are within the Classpath
+     * This is the best way to determine if a specific plugin exists and will be
+     * loaded. If the plugin package isn't loaded, we shouldn't bother waiting
+     * for it!
+     * @param packages String Array of package names to check
+     * @return Success or Failure
+     */
+    private static boolean packagesExists(String...packages) {
+        try {
+            for (String pkg : packages) {
+                Class.forName(pkg);
+            }
+            return true;
+        } catch (Exception e) {
+            return false;
+        }
+    }
+
+    public double updateCheck(double currentVersion) throws Exception {
+        String pluginUrlString = "http://dev.bukkit.org/server-mods/vault/files.rss";
+        try {
+            URL url = new URL(pluginUrlString);
+            Document doc = DocumentBuilderFactory.newInstance().newDocumentBuilder().parse(url.openConnection().getInputStream());
+            doc.getDocumentElement().normalize();
+            NodeList nodes = doc.getElementsByTagName("item");
+            Node firstNode = nodes.item(0);
+            if (firstNode.getNodeType() == 1) {
+                Element firstElement = (Element)firstNode;
+                NodeList firstElementTagName = firstElement.getElementsByTagName("title");
+                Element firstNameElement = (Element) firstElementTagName.item(0);
+                NodeList firstNodes = firstNameElement.getChildNodes();
+                return Double.valueOf(firstNodes.item(0).getNodeValue().replace("Vault", "").replaceFirst(".", "").trim());
+            }
+        }
+        catch (Exception localException) {
+        }
+        return currentVersion;
+    }
+
+    public class VaultListener implements Listener {
+
+        @EventHandler(priority = EventPriority.MONITOR)
+        public void onPlayerJoin(PlayerJoinEvent event) {
+            Player player = event.getPlayer();
+            if (getVault().getPerms().has(player, "vault.admin")) {
+                try {
+                    if (newVersion > currentVersion) {
+                        player.sendMessage(newVersion + " is out! You are running " + currentVersion);
+                        player.sendMessage("Update Vault at: http://dev.bukkit.org/server-mods/vault");
+                    }
+                } catch (Exception e) {
+                    // Ignore exceptions
+                }
+            }
+        }
+
+        @EventHandler(priority = EventPriority.MONITOR)
+        public void onPluginEnable(PluginEnableEvent event) {
+            if (event.getPlugin().getDescription().getName().equals("Register") && packagesExists("com.nijikokun.register.payment.Methods")) {
+                if (!Methods.hasMethod()) {
+                    try {
+                        Method m = Methods.class.getMethod("addMethod", Methods.class);
+                        m.setAccessible(true);
+                        m.invoke(null, "Vault", new net.milkbowl.vault.VaultEco());
+                        if (!Methods.setPreferred("Vault")) {
+                            log.info("Unable to hook register");
+                        } else {
+                            log.info("[Vault] - Successfully injected Vault methods into Register.");
+                        }
+                    } catch (SecurityException e) {
+                        log.info("Unable to hook register");
+                    } catch (NoSuchMethodException e) {
+                        log.info("Unable to hook register");
+                    } catch (IllegalArgumentException e) {
+                        log.info("Unable to hook register");
+                    } catch (IllegalAccessException e) {
+                        log.info("Unable to hook register");
+                    } catch (InvocationTargetException e) {
+                        log.info("Unable to hook register");
+                    }
+                }
+            }
+        }
+    }
+    public static Vault getVault() {
+        return vault;
+    }
+}
